chore: fix logic for running all on common change
Before, the safeguard would activate as long as the src common code
changed, even if the caller explicitly opts into certain program names.
We fix this by making that check only pass if --lazy was also provided.
In other words, the safeguard only applies when one is attempting to run
in lazy mode.